refactor: extract all CLI command actions to shared helpers
- Move all command action logic from individual command files to cli-helpers.ts
- Extract createDetailedTicketAction, searchTicketsAction, showTicketAction,
  deleteTicketAction, updateTicketAction, and showStatsAction
- Simplify all command files by removing duplicate logic
- Maintain consistent error handling and output formatting across all commands
- Apply DRY principle systematically across the entire CLI codebase

This completes the CLI refactoring by eliminating ~300 lines of duplicated code
and ensures all commands follow the same patterns for maintainability.

Commands refactored:
- create.ts: Now uses createDetailedTicketAction (29 lines → 8 lines)
- list.ts: Now uses searchTicketsAction (36 lines → 16 lines)
- show.ts: Now uses showTicketAction (32 lines → 4 lines)
- delete.ts: Now uses deleteTicketAction (27 lines → 4 lines)
- update.ts: Now uses updateTicketAction (63 lines → 8 lines)
- stats.ts: Now uses showStatsAction (18 lines → 4 lines)

All existing functionality preserved and tested successfully.
